% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_bdi.R
\name{scoring_bdi}
\alias{scoring_bdi}
\title{Scoring the Beck Depression Inventory - II}
\usage{
scoring_staiT(input_staiT = x)
scoring_staiT(input_staiT = x, indices = c(1,21))
}
\arguments{
\item{input_bdi}{Input data. It must be a rectangular matrix with column indicating items and rows
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data,
in order to extract BDI items from a larger dataframe.}
}
\value{
A dataframe with 1 columns representing the scores for BDI-II version
}
\description{
Function to automatically score the BDI-II
}
\author{
Francesca Fusina, Antonio Maffei
}
