% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ders.R
\name{scoring_ders}
\alias{scoring_ders}
\title{Scoring of the Difficulties in Emotion Regulation Scale}
\usage{
scoring_ders(input_ders = x)
scoring_ders(input_ders = x, indices = c(1,36))
}
\arguments{
\item{input_ders}{Input data. It must be a rectangular matrix with column indicating items and rows
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data,
in order to extract MPQ items from a larger dataframe.}
}
\value{
A dataframe with 7 columns representing the subscales from the DERS
}
\description{
Function to automatically score the Difficulties in Emotion Regulation Scale (DERS)
}
\author{
Francesca Fusina, Antonio Maffei
}
