% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_hamilton.R
\name{scoring_hamilton}
\alias{scoring_hamilton}
\title{Function to automatically score the Hamilton Rating Scale for Depression}
\usage{
scoring_hamilton(dataframe = x, init_col = 55, end_col = 94, ID_col = 8)
}
\arguments{
\item{dataframe}{Input data. It must be a rectangular matrix with column indicating items and rows 
indicating subjects.}

\item{init_col}{Index of the first column storing the first item of the questionnaire.}

\item{end_col}{Index of the last column storing the first item of the questionnaire.}

\item{ID_col}{Index of the column storing subject ID}
}
\value{
A dataframe with Hamilton scores for each subject
}
\description{
\code{scoring_hamilton} Returns the corrected scores for HAM-D
}
\author{
Antonio Maffei
}
