% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_lsrp.R
\name{scoring_lsrp}
\alias{scoring_lsrp}
\title{Scoring of the Levenson Self-Report Psychopathy Scale}
\usage{
scoring_lsrp(input_lsrp = x)
scoring_lsrp(input_lsrp = x, indices = c(1,26))
}
\arguments{
\item{input_lsrp}{Input data. It must be a rectangular matrix with column indicating items and rows 
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data, 
in order to extract MPQ items from a larger dataframe.}
}
\value{
A dataframe with 2 columns representing the subscales from the Levenson Self Report Psychopathy Scale
}
\description{
Function to automatically score the Levenson Self-Report Psychopathy Scale
}
\author{
Antonio Maffei
}
