% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_mpq.R
\name{scoring_mpq}
\alias{scoring_mpq}
\title{Scoring of the Multidimensional Personality Questionnaire}
\usage{
scoring_mpq(input_mpq = x)
scoring_mpq(input_mpq = x, indices = c(1,155))
}
\arguments{
\item{input_mpq}{Input data. It must be a rectangular matrix with column indicating items and rows 
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data, 
in order to extract MPQ items from a larger dataframe.}
}
\value{
A dataframe with 15 columns representing the subscales from the MPQ
}
\description{
Function to automatically score the Multidimensional Personality Questionnaire
}
\author{
Antonio Maffei
}
