% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_staiS.R
\name{scoring_staiS}
\alias{scoring_staiS}
\title{Scoring the STAI-Y2 (State version)}
\usage{
scoring_staiS(input_staiS = x)
scoring_staiS(input_staiS = x, indices = c(1,20))
}
\arguments{
\item{input_staiS}{Input data. It must be a rectangular matrix with column indicating items and rows
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data,
in order to extract MPQ items from a larger dataframe.}
}
\value{
A dataframe with 1 columns representing the scores for STAI State version
}
\description{
Function to automatically score the STAI-Y2
}
\author{
Franecsca Fusina, Antonio Maffei
}
