% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_staiT.R
\name{scoring_staiT}
\alias{scoring_staiT}
\title{Scoring the STAI-Y1 (Trait version)}
\usage{
scoring_staiT(input_staiT = x)
scoring_staiT(input_staiT = x, indices = c(1,20))
}
\arguments{
\item{input_staiT}{Input data. It must be a rectangular matrix with column indicating items and rows 
indicating subjects.}

\item{indices}{Two values vector indicating the first and the last item in the input data, 
in order to extract MPQ items from a larger dataframe.}
}
\value{
A dataframe with 1 columns representing the scores for STAI Trait version
}
\description{
Function to automatically score the STAI-Y1
}
\author{
Antonio Maffei
}
